Adding support for Pushing light cycle
Code changes for basic implementation of pushing the light cycle to the arduino.

This will be going under further improvements once the arduino client has been fleshed out to support incoming signals as well.

This is needed before final development of the application can occur since we need it to load the current cycle and time from the arduino as well as to confirm that the cycle was successfully loaded into the arduino