Initial Working Implementation
BRIEF
-------
Changes have been made in order to get the light cycle from the arduino and then update the  analog clock widget according to what has been received.  This has been done by adding support to the ArduinoClient in order to not only send data but also to receive data.

OTHER MINOR IMPROVEMENTS
-----------------------------------
A Syn message is also sent at startup in order to request the current time trackers from the arduino in order to populate the clock.

Some polish was also added in the form of style sheets to the UI in order to make the application slightly more presentable

FUTURE IMPROVEMENTS
---------------------------
Currently the syn message is only sent at the start of the application, this should either be changed to be done more frequently OR the arduino should be adjusted in order to send the light cycle after its current time has been changed. This way we will be able to pick up issues such as power failure that may reset the time

Error checking should also be added to the Start and Stop times to enforce that the start time must be before the stop time

A later update should also be made in order to support full windowed support of the application although this is of a much lower priority than the other issues and features still required.